Bug fix to Kronky, and typo errors in the docs
In the Kronky game, if the cage was "fired" with a person standing too close
to the edge then it was "locked" in movement because the system thought
it was "travelling".  The bug fix involves resetting the player correctly
in this scenario so it knows it is not.

Also fixed various typos in the docs.