Re-vamp of Library to use classes, and bug fixes
A revamp to replace "function" and "prototype" style syntax for
javascript classes to use "class" and "extend".  Also make
more use of "let" and "const" semantics.

Added the "BElite" game to the examples.  This is an impelemntation
of David Braben's and Ian Bell's AcornSoft game "ELITE" for the
BBC Micro.  Some ship designs and sounds in Christian Pinder's "ELITE
NewKind" was used there too.

Various improvements to the library is included, inclusing the ability
to place the camera on any object - "lCamera.setTo(object)".

There are ome improvements to the shader documentation and code as well.

Corrections to spelling mistakes I have noticed are here as well.